Admin improvements
- Converted counts to long
- URL shows count for every status code
- Get every count from database. Cached in redis for 6 hours
- TTL naming consistency

+++ src/main/kotlin/com/blikeng/chatapp/dtos/administration/SiteInfoDTO.kt
@@ -4,7 +4,7 @@ data class SiteInfoDTO (
     val connectedUsers: Double,
     val totalSessions: Double,
     val activeRooms: Double,
-    val totalUsers: Int,
-    val totalRooms: Int,
-    val bannedUsers: Int,
+    val totalUsers: Long,
+    val totalRooms: Long,
+    val bannedUsers: Long,
 )
